requirements.txt
fastmcp>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.31.0
psycopg2-binary>=2.9.0
python-crontab>=2.7.1
tenacity>=8.2.0
python-dotenv>=1.0.0
prometheus-client>=0.20.0
cryptography>=41.0.0
boto3>=1.28.0
redis>=5.0.0
